feat(profile): promote friends-profile bottom sheet to full-page screen (#435)
Extract the body markup from ContactProfileSheet into a shared
ContactProfileBody component, then add a new ContactProfileScreen that
renders the same body inside a full-page route with a back button +
alias + npub-truncated header.

The bottom sheet stays as a sibling for any caller that prefers the
quick-glance variant; new entry points should navigate to the screen.
